% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chi2.R
\name{chi2}
\alias{chi2}
\title{Chi-Squared Statistic}
\usage{
chi2(obsn, expn)
}
\arguments{
\item{obsn}{A numeric vector of observed counts.}

\item{expn}{A numeric vector of expected counts.}
}
\value{
A list of length two.
  The first element is a numeric vector of length three:
    \code{chistat}, chi-squared statistic;
    \code{df}, degrees of freedom; and
    \code{pval}, P value.
  The second element is a numeric vector the same length as \code{obsn},
    contributions to the chi-squared.
}
\description{
Calculate the chi-squared statistic from observed and expected counts.
}
\examples{
chi2(c(10, 8, 3), c(7, 7, 7))
}
\seealso{
\code{\link{chisq.test}}.
}

