% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coefprobit}
\alias{coefprobit}
\title{Calculate the Coefficients of a Probit Regression Fit}
\usage{
coefprobit(pfit)
}
\arguments{
\item{pfit}{An object of class \code{\link{glm}} representing a probit regression fit to dose-effect data,
typically the result of a call to \code{\link{fitprobit}}.}
}
\value{
A numeric vector of length three, the intercept and slope of the dose-response curve, each with 95\% confidence limits.
}
\description{
Calculate the coefficients from a fitted probit regression model.
}
\examples{
conc <- c(0.0625, 0.125, 0.25, 0.5, 1)
numtested <- rep(8, 5)
nalive <- c(1, 4, 4, 7, 8)
mydat <- dataprep(dose=conc, ntot=numtested, nfx=nalive)
myfit <- fitprobit(mydat)
coefprobit(myfit)
}

