% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correctval}
\alias{correctval}
\title{Predict the corrected proportion using a model fit of Table 1 of Litchfield and Wilcoxon (1949)}
\usage{
correctval(val, fit)
}
\arguments{
\item{val}{A numeric vector of expected values (as proportions).}

\item{fit}{A model object that can be used to predict the corrected values (as proportions) from \code{distexpprop5},
the distance from the expected values (as proportions) and 0.5.  Typically the output from \code{\link{gamtable1}()}.}
}
\value{
A numeric vector of corrected values (as proportions), the same length as \code{val}.
}
\description{
Given an expected proportion, calculate the corrected proportion using a model fit of Table 1 of Litchfield and Wilcoxon (1949).
}
\examples{
gamfit <- gamtable1()
correctval(c(0.37, 0.5, 0.63), gamfit)
}
\references{
Litchfield, JT Jr. and F Wilcoxon.  1949.
\href{http://jpet.aspetjournals.org/content/96/2/99.abstract}{A simplified method of evaluating dose-effect experiments}.
Journal of Pharmacology and Experimental Therapeutics 99(2):99-113.
}

