% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataprep}
\alias{dataprep}
\title{Prepare Data}
\usage{
dataprep(dose, ntot, ndead)
}
\arguments{
\item{dose}{A numeric vector of chemical concentrations.}

\item{ntot}{A numeric vector of the number of individuals that were tested at each dose.}

\item{ndead}{A numeric vector of the number of individuals that died at each dose.}
}
\value{
A data frame with eight columns (ordered by dose and pdead), seven numeric vectors and one logical vector:
	dose - chemical concentrations.
	ntot - the number of individuals that were tested at each dose.
	ndead - the number of individuals that died at each dose.
	pdead - the proportion of individuals that died at each dose.
	logdose - log transformed dose, \code{log10(dose)}.
	bitpdead - probit transformed proportional mortality, \code{\link{probit}(pdead)}.
	mcateg - mortality category: 0 for none dead, 100 for all dead, and 50 for any proportional mortality.
	LWkeep - logical vector identifying records to keep for Litchfield and Wilcoxon (1949, step A1) method.
}
\description{
Prepare dose-effect data for evaluation.
}
\examples{
conc <- c(0.0625, 0.125, 0.25, 0.5, 1)
numtested <- rep(8, 5)
nmort <- c(1, 4, 4, 7, 8)
dataprep(dose=conc, ntot=numtested, ndead=nmort)
}
\references{
Litchfield, JT Jr. and F Wilcoxon.  1949.
	\href{http://jpet.aspetjournals.org/content/96/2/99.abstract}{A simplified method of evaluating dose-effect experiments}.
	Journal of Pharmacology and Experimental Therapeutics 99(2):99-113.
}

