% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitprobit}
\alias{fitprobit}
\title{Fit a Probit Regression to Dose-Effect Data}
\usage{
fitprobit(DEdata)
}
\arguments{
\item{DEdata}{A data frame of dose-effect data (typically, the output from \code{\link{dataprep}})
containing eight variables: dose, ntot, nfx, pfx, log10dose, bitpfx, fxcateg, and LWkeep.}
}
\value{
A an object of class \code{\link{glm}}.
}
\description{
Fit a probit regression to dose-effect data, using the log10 of the dose, the binomial family, and the probit link.
}
\examples{
conc <- c(0.0625, 0.125, 0.25, 0.5, 1)
numtested <- rep(8, 5)
nalive <- c(1, 4, 4, 7, 8)
mydat <- dataprep(dose=conc, ntot=numtested, nfx=nalive)
fitprobit(mydat)
}

