% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fxcat.r
\name{fxcat}
\alias{fxcat}
\title{Define Effect Category}
\usage{
fxcat(dat)
}
\arguments{
\item{dat}{A data frame of raw toxicity data, including these three variables:
dose (the concentration of the applied chemical), ntot (the number of individuals tested), and nfx (the number of affected individuals).}
}
\value{
A numeric vector the same length as \code{prob} with quantiles on the probit scale.
}
\description{
Define three effect categories, 0 for none affected, 100 for all affected, and 50 for other proportions affected.
}
\examples{
test <- data.frame(
	dose=c(0.0625, 0.125, 0.25, 0.5),
	ntot=rep(8, 4),
	nfx = c(0, 4, 6, 8))
cbind(test, fxcat(test))
}

