% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{LW1949}
\alias{LW1949}
\alias{LW1949-package}
\title{Automated Litchfield and Wilcoxon (1949) Evaluation of Dose-Effect Experiments}
\description{
\pkg{LW1949} is an automated approach to Litchfield and Wilcoxon's (1949) evaluation of dose-effect experiments.
\pkg{LW1949} was first introduced by Adams et al. (\emph{in preparation}).
}
\details{
\emph{U.S. Geological Survey} (USGS) Computer Program \pkg{LW1949} version 2014-11.
Written by Jean V. Adams, USGS - Great Lakes Science Center \href{http://www.glsc.usgs.gov/}{www.glsc.usgs.gov}, Ann Arbor, Michigan, USA.
Written in programming language R (R Core Team, 2014, www.R-project.org), version 3.1.1 (2014-07-10).
Run on a PC with Intel(R) Core(TM) I7-4600m CPU, 2.90 GHz processor, 16.0 GB RAM, and
Microsoft Windows 7 Enterprise operating system 2009 Service Pack 1.
Source code is available from Jean V. Adams on GitHub, \href{https://github.com/JVAdams/LW1949}{https://github.com/JVAdams/LW1949}, \email{jvadams@usgs.gov}.

\emph{Disclaimer:}  Although this program has been used by the USGS, no warranty, expressed or implied, is
made by the USGS or the United States Government as to the accuracy and functioning of
the program and related program material nor shall the fact of distribution constitute
any such warranty, and no responsibility is assumed by the USGS in connection therewith.
}
\references{
Adams, JV, KS Slaght, and MA Boogaard.  \emph{In preparation}.
An automated approach to Litchfield and Wilcoxon's evaluation of dose-effect experiments.

Litchfield, JT Jr. and F Wilcoxon.  1949.
A simplified method of evaluating dose-effect experiments.
Journal of Pharmacology and Experimental Therapeutics 99(2):99-113.
\href{http://jpet.aspetjournals.org/content/96/2/99.abstract}{[link]}.
}

