% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LWP}
\alias{LWP}
\title{User Friendly Evaluation of Dose-Effect Experiments using Litchfield-Wilcoxon and Probit Methods}
\usage{
LWP(rawfile = NULL, descrcolz = 4, saveplots = TRUE, showplots = FALSE,
  saveresults = TRUE, showresults = TRUE, returnresults = FALSE)
}
\arguments{
\item{rawfile}{A character scalar specifying the path of the input data as a csv file.  If NULL, default, the user will be
prompted to browse to a file using a menu.}

\item{descrcolz}{A numeric scalar, the number of columns to use as the description of the test,
from 1 to \code{descrcolz}, default 4.}

\item{saveplots}{A logical scalar indicating if plots should be saved to a pdf file, default TRUE.}

\item{showplots}{A logical scalar indicating if plots should be shown on screen, default FALSE.}

\item{saveresults}{A logical scalar indicating if results should be saved to a csv file, default TRUE.
The csv file is given the same name (plus the suffix "Smry") and is placed in the same directory as the input file.}

\item{showresults}{A logical scalar indicating if results should be printed to the console, default TRUE.}

\item{returnresults}{A logical scalar indicating if results should be returned by the funcion, default FALSE.}
}
\value{
If \code{returnresults=TRUE}, a data frame with 11 rows per test and 2 more columns than the input data.
	Three columns from the input data are not included (\code{TFM Conc. (mg/L)}, \code{No. Tested}, and \code{No. Dead}).
	Five columns are added: the parameter (\code{param}), the method used (\code{method}),
	the estimate (\code{estimate}), and the 95% confidence interval of the estimate (\code{lower95ci} and \code{upper95ci})
}
\description{
User friendly evaluation of dose-effect experiments using automated Litchfield Wilcoxon (1949) and probit estimation methods.
	This function has been tailored for non-R users with input data set up in a particular way (see Details).
}
\details{
The input data must include at least these seven columns, with these names:
"Test ID" = A character or numeric vector, the unique identifier for each test
"Source" = A character vector, the source of the chemical
"Batch" = A character or numeric vector, the batch of the chemical
"Species" = A character vector, the species tested
"TFM Conc. (mg/L)" = A numeric vector, the concentration of TFM in mg/L
"No. Tested" = A numeric vector, the number of animals tested
"No. Dead" = A numeric vector, the number of animals dead
}
\examples{
\dontrun{
LWP()
}
}

