% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcat}
\alias{mcat}
\title{Define Mortality Category}
\usage{
mcat(dat)
}
\arguments{
\item{dat}{A data frame of raw toxicity data, including these three variables:
dose (the concentration of the applied chemical), ntot (the number of individuals tested), and ndead (the number of dead individuals).}
}
\value{
A numeric vector the same length as \code{prob} with quantiles on the probit scale.
}
\description{
Define three mortality categories, 0 for nonoe dead, 100 for all dead, and 50 for anything in between, partial mortality.
}
\examples{
test <- data.frame(
	dose=c(0.0625, 0.125, 0.25, 0.5),
	ntot=rep(8, 4),
	ndead = c(0, 4, 6, 8))
cbind(test, mcat(test))
}

