% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predlinear}
\alias{predlinear}
\title{Determine the Effective Dose from a Linear Regression Fit}
\usage{
predlinear(pct, b0, b1)
}
\arguments{
\item{pct}{A numeric vector of effects (in percents) for which to estimate the effect dose(s).}

\item{b0}{A numeric vector (more commonly a scalar) giving the intercept of the dose-response curve.}

\item{b1}{A numeric vector (more commonly a scalar) giving the slope of the dose-response curve.
x = dose (the concentration of the applied chemical on the log10 scale),
and y, the proportion of affected individuals (on the probit scale, with 0s converted to 0.1\% and 1s converted to 99.9\%).}
}
\value{
A numeric vector the same length as \code{pct} giving the estimated dose at the specified percent effect.
}
\description{
Determine the effective dose for a specified percent effect from the intercept and slope of a linear regression.
}
\examples{
predlinear(c(16, 50, 84, 99.9), 1.700875, 2.199559)
}

