% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predLWCI}
\alias{predLWCI}
\title{Estimate the 95\% Confidence Interval of an Effective Dose from a Linear Regression Fit}
\usage{
predLWCI(Y, EDY, fED50, fS)
}
\arguments{
\item{Y}{A numeric vector of effects (in percents) for which the effective dose confidence intervals are to be calculated.}

\item{EDY}{A numeric vector of effective dose(s) corresponding to \code{Y}.}

\item{fED50}{A numeric vector (more commonly a scalar) giving the expansion factor(s) for the ED50
of the dose-response curve (typically, output from \code{\link{fitLW}}.}

\item{fS}{A numeric vector (more commonly a scalar) giving the expansion factor(s) for the Litchfield and Wilcoxon (1949)
slope (typically, output from \code{\link{fitLW}}.}
}
\value{
An n*2 numeric matrix with the Litchfield and Wilcoxon (1949) lower and upper 95\% confidence intervals
	of the effective dose, where n is the length of \code{Y} and \code{EDY}.
}
\description{
Estimate the 95\% confidence interval of an effective dose for a specified percent effect
	from the intercept and slope of a linear regression using the approach of Litchfield and Wilcoxon (1949).
}
\details{
Follows methods outlined in Litchfield and Wilcoxon (1949) on page 105, using equation 13 in the Appendix
	instead of Nomograph 4.
}
\examples{
dose <- c(0.0625, 0.125, 0.25, 0.5, 1)
ntested <- rep(8, 5)
nalive <- c(1, 4, 4, 7, 8)
mydat <- dataprep(dose=dose, ntot=ntested, nfx=nalive)
mydat
flw <- fitLW(mydat)
pcts <- c(25, 99.9)
EDYs <- predlinear(pcts, b0=flw$params[1], b1=flw$params[2])
predLWCI(Y=pcts, EDY=EDYs, fED50=flw$LWest["fED50"], fS=flw$LWest["fS"])
}
\references{
Litchfield, JT Jr. and F Wilcoxon.  1949.
	A simplified method of evaluating dose-effect experiments.
	Journal of Pharmacology and Experimental Therapeutics 99(2):99-113.
	\href{http://jpet.aspetjournals.org/content/96/2/99.abstract}{[link]}.
}

