% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predprobit.R
\name{predprobit}
\alias{predprobit}
\title{Determine the Effective Dose from a Probit Regression Fit}
\usage{
predprobit(pct, pfit)
}
\arguments{
\item{pct}{A numeric scalar of the effect (as a percent) for which to estimate the
effective dose.}

\item{pfit}{An object of class \code{\link{glm}} representing a probit regression fit
to dose-effect data, typically the result of a call to
\code{\link{fitprobit}}.}
}
\value{
A numeric vector of length three, the effective dose and the lower and
  upper 95\% confidence limits.
}
\description{
Determine the effective dose for a specified percent effect from a fitted
  probit regression model.
}
\examples{
conc <- c(0.0625, 0.125, 0.25, 0.5, 1)
numtested <- rep(8, 5)
nalive <- c(1, 4, 4, 7, 8)
mydat <- dataprep(dose=conc, ntot=numtested, nfx=nalive)
myfit <- fitprobit(mydat)
predprobit(50, myfit)
}

