% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/probit.r
\name{probit}
\alias{probit}
\title{Convert Proportions to the Probit Scale}
\usage{
probit(prob)
}
\arguments{
\item{prob}{A numeric vector of proportions.}
}
\value{
A numeric vector the same length as \code{prob} with quantiles on the probit scale.
}
\description{
Convert proportions to the probit scale.
}
\details{
Simply calls \code{\link{qnorm}(prob)}.
}
\examples{
probit(c(0.001, 0.01, 0.1, 0.5, 0.9, 0.99, 0.999))
}

