% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chi2}
\alias{chi2}
\title{Chi-Squared Statistic}
\usage{
chi2(obsn, expn)
}
\arguments{
\item{obsn}{A numeric vector of observed counts.}

\item{expn}{A numeric vector of expected counts.}
}
\value{
A list with three elements: chistat, a numeric scalar, the chi-squared statistic;
	pval, a numeric scalar, its associated P value; and df, an integer, the degrees of freedom of \code{chistat}.
}
\description{
Calculate the chi-squared statistic from observed and expected counts.
}
\examples{
chi2(c(10, 8, 3), c(7, 7, 7))
}
\seealso{
\code{\link{chisq.test}}.
}

