% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chi2}
\alias{chi2}
\title{Chi-Squared Statistic}
\usage{
chi2(obsn, expn)
}
\arguments{
\item{obsn}{A numeric vector of observed counts.}

\item{expn}{A numeric vector of expected counts.}
}
\value{
A list of length two.  The first element is a numeric vector of length three: chistat, the chi-squared statistic;
	df, the degrees of freedom of \code{chistat}; and pval, their associated P value.
	The second element is a numeric vector the same length as \code{obsn}, contributions to the chi-squared.
}
\description{
Calculate the chi-squared statistic from observed and expected counts.
}
\examples{
chi2(c(10, 8, 3), c(7, 7, 7))
}
\seealso{
\code{\link{chisq.test}}.
}

