% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LWauto}
\alias{LWauto}
\title{Apply Litchfield and Wilcoxon Evaluation of Dose-Effect Experiments}
\usage{
LWauto(DEdata)
}
\arguments{
\item{DEdata}{A data frame of dose-effect data (typically, the output from \code{\link{dataprep}}
containing at eight variables: dose, ntot, nfx, pfx, logdose, bitpfx, fxcateg, and LWkeep.}
}
\value{
A list of length three:
 chi = the chi-squared statistic with associated P value and degrees of freedom,
 params = the estimated LC50 and LC99.9 on the log10 scale used to describe the dose-response curve,
 LWest = the Litchfield Wilcoxon estimates of LC50 with 95% confidence intervals and other metrics used in their step-by-step approach
 (LC16, LC84, S, and slope).
}
\description{
Automatically apply Litchfield and Wilcoxon's (1949) evaluation of dose-effect experiments.
}
\examples{
dose <- c(0.0625, 0.125, 0.25, 0.5, 1)
numtested <- rep(8, 5)
numalive <- c(1, 4, 4, 7, 8)
mydat <- dataprep(dose=dose, ntot=numtested, nfx=numalive)
mydat
LWauto(mydat)
}

